Minor fixes
This update makes lead bloons not pop when travelling over road spikes. It also disables MOAB-class bloons, as they are not yet functional and cause crashing. I fixed the interface for the upgrader, but there is so far no actual functionality in them. This commit is WAY behind the one on my regular machine, but this should do for now.